windows support + minor refactoring

+++ src/services/daprClient.ts
@@ -4,9 +4,11 @@
 import * as url from 'url';
 import * as vscode from 'vscode';
 import * as nls from 'vscode-nls';
+import * as os from 'os';
 import { DaprApplication } from "./daprApplicationProvider";
 import { HttpClient, HttpResponse } from './httpClient';
 import { getLocalizationPathForFile } from '../util/localization';
+import { Process } from '../util/process';
 
 
 const localize = nls.loadMessageBundle(getLocalizationPathForFile(__filename));
@@ -64,6 +66,12 @@ export default class HttpDaprClient implements DaprClient {
     }
 
     stopApp(application: DaprApplication): void {
-        process.kill(application.pid, 'SIGKILL')
+        if (os.platform() === 'win32') {
+            // NOTE: Windows does not support SIGTERM/SIGINT/SIGBREAK, so there can be no graceful process shutdown.
+            //       As a partial mitigation, use `taskkill` to kill the entire process tree.
+            void Process.exec(`taskkill /pid ${application.pid} /t /f`);
+        } else {
+            process.kill(application.pid, 'SIGKILL')
+        }
     }
 }
